Support MasterCook txt

+++ src/core/integrations/apps/bigoven.rs
@@ -125,7 +125,7 @@ impl From<BigOvenRecipe> for RecipeSchema {
             aggregate_rating: if r.taste_rating > 0 {
                 Some(AggregateRating {
                     at_type: AtType::AggregateRating,
-                    rating_value: Some(NumberOrText::Number(r.taste_rating as i64)),
+                    rating_value: Some(NumberOrText::Number(r.taste_rating as f64)),
                     ..Default::default()
                 })
             } else {
